+++ src/circuit/states.rs
@@ -21,7 +21,7 @@
 
 use crate::Complex;
 use crate::QuantrError;
-use crate::{complex_Re, complex_zero};
+use crate::{complex_Re, COMPLEX_ZERO};
 use std::collections::HashMap;
 use std::hash::Hash;
 
@@ -35,8 +35,7 @@ pub enum Qubit {
 }
 
 impl Qubit {
-    // Change to `kronecker_prod`
-    /// Defines the tensor product on two qubits.
+    /// Defines the kronecker product on two qubits.
     ///
     /// # Example
     /// ```
@@ -45,16 +44,16 @@ impl Qubit {
     /// let qubit_a: Qubit = Qubit::Zero; // |0>
     /// let qubit_b: Qubit = Qubit::One;  // |1>
     ///
-    /// let new_product: ProductState = qubit_a.join(qubit_b); // |0> ⊗ |1> = |01>
-    /// assert_eq!(new_product.state.as_slice(), &[Qubit::Zero, Qubit::One])
+    /// let new_product: ProductState = qubit_a.kronecker_prod(qubit_b); // |0> ⊗ |1> = |01>
+    /// assert_eq!(new_product.qubits.as_slice(), &[Qubit::Zero, Qubit::One])
     /// ```
-    pub fn join(self, other: Qubit) -> ProductState {
+    pub fn kronecker_prod(self, other: Qubit) -> ProductState {
         ProductState::new(&[self, other])
     }
 
     // Deprecated as it will soon change to `into_state`.
     /// Converts the [Qubit] to a [ProductState] struct.
-    pub fn as_state(self) -> ProductState {
+    pub fn into_state(self) -> ProductState {
         ProductState::new(&[self])
     }
 }
@@ -67,7 +66,7 @@ impl Qubit {
 pub struct ProductState {
     /// Each element of `state` is mapped to bra-ket notation like so:
     /// `Vec<Qubit>{a, b, ..., c} -> |a...bc>`
-    pub state: Vec<Qubit>,
+    pub qubits: Vec<Qubit>,
 }
 
 impl ProductState {
@@ -79,7 +78,7 @@ impl ProductState {
     /// `&[a, b, ..., c] -> |ab...c>`
     pub fn new(product_state: &[Qubit]) -> ProductState {
         ProductState {
-            state: product_state.to_vec(),
+            qubits: product_state.to_vec(),
         }
     }
 
@@ -88,7 +87,7 @@ impl ProductState {
     // Changes the qubits at specified positions within the product state with a slice of other
     // qubits.
     pub(super) fn insert_qubits(&self, qubits: &[Qubit], pos: &[usize]) -> ProductState {
-        let mut edited_qubits: Vec<Qubit> = self.state.clone();
+        let mut edited_qubits: Vec<Qubit> = self.qubits.clone();
         let num_qubits: usize = qubits.len();
 
         if num_qubits != pos.len() {
@@ -103,7 +102,7 @@ impl ProductState {
 
     // Returns the dimension of the product state.
     fn num_qubits(&self) -> usize {
-        self.state.len()
+        self.qubits.len()
     }
 
     /// Inverts a binary digit that represents the product state.
@@ -115,8 +114,8 @@ impl ProductState {
             return Err(QuantrError { message: format!("The position of the binary digit, {}, is out of bounds. The product dimension is {}, and so the position must be strictly less.", place_num, self.num_qubits()) });
         }
 
-        let old_qubit: Qubit = self.state[place_num].clone();
-        self.state[place_num] = if old_qubit == Qubit::Zero {
+        let old_qubit: Qubit = self.qubits[place_num].clone();
+        self.qubits[place_num] = if old_qubit == Qubit::Zero {
             Qubit::One
         } else {
             Qubit::Zero
@@ -128,20 +127,19 @@ impl ProductState {
     /// Concatenate a product state with a qubit.
     ///
     /// In effect, this is using the tensor prodcut to create a new state.
-    pub fn join(mut self, other: Qubit) -> ProductState {
-        self.state.push(other);
+    pub fn kronecker_prod(mut self, other: Qubit) -> ProductState {
+        self.qubits.push(other);
         self
     }
 
     // Returns the qubit in the product state given a position.
     pub(super) fn get(&self, qubit_number: usize) -> Qubit {
-        self.state[qubit_number]
+        self.qubits[qubit_number]
     }
 
-    // `into_string`
     /// Returns the labelling of the product state as a String.
-    pub fn as_string(&self) -> String {
-        self.state
+    pub fn to_string(&self) -> String {
+        self.qubits
             .iter()
             .map(|q| match q {
                 Qubit::Zero => "0",
@@ -150,16 +148,15 @@ impl ProductState {
             .collect::<String>()
     }
 
-    // Change to `into_super_position`
     /// Returns the [ProductState] as a [SuperPosition].
-    pub fn to_super_position(self) -> SuperPosition {
+    pub fn into_super_position(self) -> SuperPosition {
         SuperPosition::new(self.num_qubits())
             .set_amplitudes_from_states_unchecked(&HashMap::from([(self, complex_Re!(1f64))]))
     }
 
     // Converts the computational basis labelling (a binary integer), into base 10.
     fn comp_basis(&self) -> usize {
-        self.state
+        self.qubits
             .iter()
             .rev()
             .enumerate()
@@ -201,7 +198,7 @@ pub struct SuperPosition {
 /// # Example
 /// ```
 /// use quantr::states::{ProductState, Qubit, SuperPosition};
-/// use quantr::{complex_Re, complex_Re_vec, complex_zero, Complex};
+/// use quantr::{complex_Re, complex_Re_vec, COMPLEX_ZERO, Complex};
 /// use std::f64::consts::FRAC_1_SQRT_2;
 ///
 /// let super_pos: SuperPosition = SuperPosition::new(2)
@@ -210,13 +207,13 @@ pub struct SuperPosition {
 /// let mut iterator_super_pos = super_pos.into_iter();
 ///
 /// assert_eq!(iterator_super_pos.next(),
-///     Some((ProductState::new(&[Qubit::Zero, Qubit::Zero]), complex_zero!())));
+///     Some((ProductState::new(&[Qubit::Zero, Qubit::Zero]), COMPLEX_ZERO)));
 /// assert_eq!(iterator_super_pos.next(),
 ///     Some((ProductState::new(&[Qubit::Zero, Qubit::One]), complex_Re!(FRAC_1_SQRT_2))));
 /// assert_eq!(iterator_super_pos.next(),
 ///     Some((ProductState::new(&[Qubit::One, Qubit::Zero]), complex_Re!(FRAC_1_SQRT_2))));
 /// assert_eq!(iterator_super_pos.next(),
-///     Some((ProductState::new(&[Qubit::One, Qubit::One]), complex_zero!())));
+///     Some((ProductState::new(&[Qubit::One, Qubit::One]), COMPLEX_ZERO)));
 /// assert_eq!(iterator_super_pos.next(), None);
 /// ```
 pub struct SuperPositionIterator<'a> {
@@ -259,7 +256,7 @@ impl SuperPosition {
 
     /// Creates a superposition in the |0..0> state.
     pub fn new(num_qubits: usize) -> SuperPosition {
-        let mut new_amps: Vec<Complex<f64>> = vec![complex_zero!(); 2usize.pow(num_qubits as u32)];
+        let mut new_amps: Vec<Complex<f64>> = vec![COMPLEX_ZERO; 2usize.pow(num_qubits as u32)];
         new_amps[0] = complex_Re!(1f64);
         SuperPosition {
             amplitudes: new_amps,
@@ -284,8 +281,8 @@ impl SuperPosition {
         &self,
         prod_state: ProductState,
     ) -> Result<Complex<f64>, QuantrError> {
-        if 2usize.pow(prod_state.state.len() as u32) != self.amplitudes.len() {
-            return Err(QuantrError { message: format!("Unable to retreive product state, |{:?}> with dimension {}. The superposition is a linear combination of states with different dimension. These dimensions should be equal.", prod_state.as_string(), prod_state.num_qubits()),});
+        if 2usize.pow(prod_state.qubits.len() as u32) != self.amplitudes.len() {
+            return Err(QuantrError { message: format!("Unable to retreive product state, |{:?}> with dimension {}. The superposition is a linear combination of states with different dimension. These dimensions should be equal.", prod_state.to_string(), prod_state.num_qubits()),});
         }
         Ok(*self.amplitudes.get(prod_state.comp_basis()).unwrap())
     }
@@ -357,7 +354,7 @@ impl SuperPosition {
         let mut total_amplitude: f64 = 0f64;
         for (states, amplitude) in amplitudes {
             if states.num_qubits() != product_size {
-                return Err(QuantrError { message: format!("The first state has product dimension of {}, whilst the state, |{}>, found as a key in the HashMap has dimension {}.", product_size, states.as_string(), states.num_qubits()) });
+                return Err(QuantrError { message: format!("The first state has product dimension of {}, whilst the state, |{}>, found as a key in the HashMap has dimension {}.", product_size, states.to_string(), states.num_qubits()) });
             }
             total_amplitude += amplitude.abs_square();
         }
@@ -368,8 +365,7 @@ impl SuperPosition {
 
         // Start conversion
 
-        let mut new_amps: Vec<Complex<f64>> =
-            vec![complex_zero!(); 2usize.pow(product_size as u32)];
+        let mut new_amps: Vec<Complex<f64>> = vec![COMPLEX_ZERO; 2usize.pow(product_size as u32)];
 
         Self::from_hash_to_array(amplitudes, &mut new_amps);
 
@@ -387,8 +383,7 @@ impl SuperPosition {
         amplitudes: &HashMap<ProductState, Complex<f64>>,
     ) -> SuperPosition {
         let product_size: usize = amplitudes.keys().next().unwrap().num_qubits();
-        let mut new_amps: Vec<Complex<f64>> =
-            vec![complex_zero!(); 2usize.pow(product_size as u32)];
+        let mut new_amps: Vec<Complex<f64>> = vec![COMPLEX_ZERO; 2usize.pow(product_size as u32)];
 
         Self::from_hash_to_array(amplitudes, &mut new_amps);
 
@@ -402,7 +397,7 @@ impl SuperPosition {
     /// Creates a HashMap of the superposition with [ProductState] as keys.
     ///
     /// The HashMap will not include states with amplitudes that are near zero.
-    pub fn as_hash_map(&self) -> HashMap<ProductState, Complex<f64>> {
+    pub fn to_hash_map(&self) -> HashMap<ProductState, Complex<f64>> {
         let mut super_pos_as_hash: HashMap<ProductState, Complex<f64>> = Default::default();
         for (i, amp) in self.amplitudes.iter().enumerate() {
             if !Self::equal_within_error(amp.abs_square(), 0f64) {
@@ -431,14 +426,9 @@ mod tests {
     #[test]
     fn converts_productstate_to_superpos() {
         assert_eq!(
-            ProductState::new(&[Qubit::One, Qubit::Zero]).to_super_position(),
+            ProductState::new(&[Qubit::One, Qubit::Zero]).into_super_position(),
             SuperPosition::new(2)
-                .set_amplitudes(&[
-                    complex_zero!(),
-                    complex_zero!(),
-                    complex_Re!(1f64),
-                    complex_zero!()
-                ])
+                .set_amplitudes(&[COMPLEX_ZERO, COMPLEX_ZERO, complex_Re!(1f64), COMPLEX_ZERO])
                 .unwrap()
         )
     }
@@ -468,10 +458,10 @@ mod tests {
         assert_eq!(
             SuperPosition::new(2)
                 .set_amplitudes(&[
-                    complex_zero!(),
+                    COMPLEX_ZERO,
                     complex_Re!(FRAC_1_SQRT_2),
                     complex_Im!(-FRAC_1_SQRT_2),
-                    complex_zero!()
+                    COMPLEX_ZERO
                 ])
                 .unwrap()
                 .get_amplitude_from_state(ProductState::new(&[Qubit::Zero, Qubit::One]))
@@ -485,10 +475,10 @@ mod tests {
         assert_eq!(
             SuperPosition::new(2)
                 .set_amplitudes(&[
-                    complex_zero!(),
+                    COMPLEX_ZERO,
                     complex_Re!(FRAC_1_SQRT_2),
                     complex_Im!(-FRAC_1_SQRT_2),
-                    complex_zero!()
+                    COMPLEX_ZERO
                 ])
                 .unwrap()
                 .get_amplitude(2)
@@ -500,10 +490,10 @@ mod tests {
     #[test]
     fn insert_qubits_in_state() {
         assert_eq!(
-            ProductState::new(&[Qubit::Zero, Qubit::Zero, Qubit::One]).state,
+            ProductState::new(&[Qubit::Zero, Qubit::Zero, Qubit::One]).qubits,
             ProductState::new(&[Qubit::One, Qubit::One, Qubit::One])
                 .insert_qubits(&[Qubit::Zero, Qubit::Zero], &[0, 1])
-                .state
+                .qubits
         );
     }
 
@@ -523,10 +513,10 @@ mod tests {
         assert_eq!(
             SuperPosition::new(2)
                 .set_amplitudes(&[
-                    complex_zero!(),
+                    COMPLEX_ZERO,
                     complex_Re!(FRAC_1_SQRT_2),
                     complex_Im!(-FRAC_1_SQRT_2),
-                    complex_zero!()
+                    COMPLEX_ZERO
                 ])
                 .unwrap()
                 .amplitudes,
@@ -579,10 +569,10 @@ mod tests {
     fn catches_retrieve_amplitude_from_wrong_state() {
         SuperPosition::new(2)
             .set_amplitudes(&[
-                complex_zero!(),
+                COMPLEX_ZERO,
                 complex_Re!(FRAC_1_SQRT_2),
                 complex_Im!(-FRAC_1_SQRT_2),
-                complex_zero!(),
+                COMPLEX_ZERO,
             ])
             .unwrap()
             .get_amplitude_from_state(ProductState::new(&[Qubit::Zero, Qubit::One, Qubit::One]))
@@ -594,10 +584,10 @@ mod tests {
     fn catches_retrieve_amplitude_from_wrong_list_pos() {
         SuperPosition::new(2)
             .set_amplitudes(&[
-                complex_zero!(),
+                COMPLEX_ZERO,
                 complex_Re!(FRAC_1_SQRT_2),
                 complex_Im!(-FRAC_1_SQRT_2),
-                complex_zero!(),
+                COMPLEX_ZERO,
             ])
             .unwrap()
             .get_amplitude(4)
@@ -609,9 +599,9 @@ mod tests {
     fn catches_super_position_breaking_conservation() {
         SuperPosition::new(2)
             .set_amplitudes(&[
-                complex_zero!(),
+                COMPLEX_ZERO,
                 complex_Re!(0.5f64),
-                complex_zero!(),
+                COMPLEX_ZERO,
                 complex_Im!(-0.5f64),
             ])
             .unwrap();