+++ tests/grovers.rs
@@ -19,12 +19,11 @@ use std::f64::consts::FRAC_1_SQRT_2;
 const ERROR_MARGIN: f64 = 0.00000001f64;
 
 #[test]
-fn grovers_3qubit() -> Result<(), QuantrError>{
+fn grovers_3qubit() -> Result<(), QuantrError> {
     let mut circuit = Circuit::new(3)?;
 
     // Kick state into superposition of equal weights
-    circuit
-        .add_repeating_gate(StandardGate::H, &[0, 1, 2])?;
+    circuit.add_repeating_gate(StandardGate::H, &[0, 1, 2])?;
 
     // Oracle
     circuit.add_gate(StandardGate::CZ(1), 2)?;
@@ -60,7 +59,7 @@ fn grovers_3qubit() -> Result<(), QuantrError>{
 
     if let Observable(bin_count) = circuit.repeat_measurement(500).unwrap() {
         for (state, count) in bin_count {
-            match state.as_string().as_str() {
+            match state.to_string().as_str() {
                 "011" | "111" => assert!(count > 200usize),
                 _ => assert_eq!(count, 0usize),
             }
@@ -115,7 +114,10 @@ fn x3sudoko() -> Result<(), QuantrError> {
     qc.add_repeating_gate(StandardGate::H, &[0, 1, 2, 3, 4, 5])?
         .add_repeating_gate(StandardGate::X, &[0, 1, 2, 3, 4, 5])?
         .add_gate(StandardGate::H, 5)?
-        .add_gate(StandardGate::Custom(cccccnot, &[0, 1, 2, 3, 4], "X".to_string()),5,)?
+        .add_gate(
+            StandardGate::Custom(cccccnot, &[0, 1, 2, 3, 4], "X".to_string()),
+            5,
+        )?
         .add_gate(StandardGate::H, 5)?
         .add_repeating_gate(StandardGate::X, &[0, 1, 2, 3, 4, 5])?
         .add_repeating_gate(StandardGate::H, &[0, 1, 2, 3, 4, 5])?;
@@ -125,7 +127,7 @@ fn x3sudoko() -> Result<(), QuantrError> {
 
     if let Observable(bin_count) = qc.repeat_measurement(5000).unwrap() {
         for (state, count) in bin_count {
-            match &state.as_string()[0..=5] {
+            match &state.to_string()[0..=5] {
                 "001100" | "001010" | "010100" | "010001" | "100010" | "100001" => {
                     assert!(count > 150usize)
                 }
@@ -139,47 +141,44 @@ fn x3sudoko() -> Result<(), QuantrError> {
 
 fn cccnot(input_state: ProductState) -> SuperPosition {
     let mut copy_state = input_state.clone();
-    if copy_state.state == [Qubit::One; CCC_NUMBER] {
-        copy_state.state[CCC_NUMBER - 1] = Qubit::Zero;
-        return copy_state.to_super_position();
-    } else if copy_state.state == {
+    if copy_state.qubits == [Qubit::One; CCC_NUMBER] {
+        copy_state.qubits[CCC_NUMBER - 1] = Qubit::Zero;
+        return copy_state.into_super_position();
+    } else if copy_state.qubits == {
         let mut temp = [Qubit::One; CCC_NUMBER];
         temp[CCC_NUMBER - 1] = Qubit::Zero;
         temp
     } {
-        copy_state.state[CCC_NUMBER - 1] = Qubit::One;
-        return copy_state.to_super_position();
+        copy_state.qubits[CCC_NUMBER - 1] = Qubit::One;
+        return copy_state.into_super_position();
     } else {
-        copy_state.to_super_position()
+        copy_state.into_super_position()
     }
 }
 
 // Implementation of a 5 controlled toffoli gate
 fn cccccnot(input_state: ProductState) -> SuperPosition {
     let mut copy_state = input_state.clone();
-    if copy_state.state == [Qubit::One; CCCCC_NUMBER] {
-        copy_state.state[CCCCC_NUMBER - 1] = Qubit::Zero;
-        return copy_state.to_super_position();
-    } else if copy_state.state == {
+    if copy_state.qubits == [Qubit::One; CCCCC_NUMBER] {
+        copy_state.qubits[CCCCC_NUMBER - 1] = Qubit::Zero;
+        return copy_state.into_super_position();
+    } else if copy_state.qubits == {
         let mut temp = [Qubit::One; CCCCC_NUMBER];
         temp[CCCCC_NUMBER - 1] = Qubit::Zero;
         temp
     } {
-        copy_state.state[CCCCC_NUMBER - 1] = Qubit::One;
-        return copy_state.to_super_position();
+        copy_state.qubits[CCCCC_NUMBER - 1] = Qubit::One;
+        return copy_state.into_super_position();
     } else {
-        copy_state.to_super_position()
+        copy_state.into_super_position()
     }
 }
 
 fn compare_complex_lists_and_register(correct_list: &[Complex<f64>], register: &SuperPosition) {
     for (i, &comp_num) in register.amplitudes.iter().enumerate() {
         // Make sure that it turns up complex
-        assert!(equal_within_error(comp_num.real, correct_list[i].real));
-        assert!(equal_within_error(
-            comp_num.imaginary,
-            correct_list[i].imaginary
-        ));
+        assert!(equal_within_error(comp_num.re, correct_list[i].re));
+        assert!(equal_within_error(comp_num.im, correct_list[i].im));
     }
 }